requirements.txt
langgraph
langchain
langchain-openai
langchain-community
streamlit
faiss-cpu
python-dotenv
PyPDF2
PyMuPDF
pytesseract
pillow
unstructured
pdfminer.six